requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28
sqlalchemy>=2.0.0
streamlit>=1.30.0
plotly>=5.17.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
pytz>=2024.1
